Renamed 'rtt_measurement' to verloc to avoid confusion
