+++ common/mixnode-common/src/lib.rs
@@ -13,4 +13,4 @@
 // limitations under the License.
 
 pub mod cached_packet_processor;
-pub mod rtt_measurement;
+pub mod verloc;
