+++ common/mixnode-common/src/verloc/listener.rs
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::rtt_measurement::error::RttError;
-use crate::rtt_measurement::packet::{EchoPacket, ReplyPacket};
+use crate::verloc::error::RttError;
+use crate::verloc::packet::{EchoPacket, ReplyPacket};
 use bytes::{BufMut, BytesMut};
 use crypto::asymmetric::identity;
 use futures::StreamExt;
